Push more job index logic from service into manager layer.
